+++ CHANGELOG.md
@@ -10,7 +10,9 @@
 - Upgrade our language baseline to ES5 ([#174](https://github.com/saneyuki/option-t.js/pull/174))
   - If you use this package for a product targeted to ES3 environment (e.g. ~IE8),
     You need to care about it.
-
+- Rename: ([#175](https://github.com/saneyuki/option-t.js/pull/175))
+  - `Maybe::isUndefinedOrNull()` -> `Maybe::isNullOrUndefined()`
+  - `Maybe::isSomeActual()` -> `Maybe::isSomeValue()`
 
 ## 5.0.0